This change allows users to define a specific deployment unit on the local file system rather than just the directory.
I.e. This means that you can configure target, so everything underneath of the maven target directory is picked up, but you can now define target/myapp.war as well and it will be picked up. Versions without that change simply showed the user an [OPENSHIFT] null message which was not really helpful.

I have added a Test Class and PowerMock, because I needed to test a private method.